Added startup Window + functions
Added startup window before launch of the app, when JDK is choosen it creates a new project with that JDK and stores it on the users desktop for easy access. Add, remove and change JDK is implemented in the startup window as well as auto detecting downloaded JDK when importing from computer.

Co-Authored-By: Munther23 <143321719+Munther23@users.noreply.github.com>